requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.4.0
drf-spectacular==0.27.2
Pillow==10.4.0

# Optional but recommended
celery==5.4.0
redis==5.0.8
drf-spectacular==0.27.2